+++ src/script.js
@@ -1,10 +1,16 @@
 const colorPickerBtn = document.getElementById("color-picker");
+const getAllBtn = document.getElementById("get-all-colors");
 const colorList = document.querySelector(".all-colors");
+const colorsAll = document.querySelector(".all-website-colors-list");
+const colorsAllTitle = document.querySelector(".all-website-colors");
 const clearAll = document.querySelector(".clear-all");
+const copiedHeader = document.querySelector(".copied");
 const pickedColors = JSON.parse(localStorage.getItem("picked-colors") || "[]");
 const pickedColorsRGB = JSON.parse(
   localStorage.getItem("picked-colors-rgb") || "[]"
 );
+
+//HEX to RGB Calculation
 
 function hexToRGB(h) {
   let r = 0,
@@ -24,11 +30,21 @@ function hexToRGB(h) {
   return "rgb(" + +r + "," + +g + "," + +b + ")";
 }
 
+//Copy Color from Text to Clipboard
+
 const copyColor = (elem) => {
   navigator.clipboard.writeText(elem.dataset.color);
   elem.innerText = "Copied";
   setTimeout(() => (elem.innerText = elem.dataset.color), 1000);
 };
+
+//Copy Color from Rect to Clipboard
+
+const copyColorRect = (elem) => {
+  navigator.clipboard.writeText(elem.dataset.color);
+};
+
+//Show selected color
 
 const showColors = () => {
   if (!pickedColors.length) return;
@@ -62,6 +78,8 @@ const showColors = () => {
 };
 showColors();
 
+//Activate EyeDropper
+
 const activateEyeDropper = () => {
   document.body.style.display = "none";
   setTimeout(async () => {
@@ -87,6 +105,8 @@ const activateEyeDropper = () => {
   }, 10);
 };
 
+//Delete all Colors stored in Local Storage
+
 const clearAllColors = () => {
   pickedColors.length = 0;
   localStorage.setItem("picked-colors", JSON.stringify(pickedColors));
@@ -95,3 +115,78 @@ const clearAllColors = () => {
 
 clearAll.addEventListener("click", clearAllColors);
 colorPickerBtn.addEventListener("click", activateEyeDropper);
+
+document.addEventListener("DOMContentLoaded", function () {
+  chrome.tabs.query({ active: true, currentWindow: true }, function (tabs) {
+    chrome.scripting.executeScript(
+      {
+        target: { tabId: tabs[0].id },
+        function: extractColors,
+      },
+      function (results) {
+        displayColors(results[0].result);
+      }
+    );
+  });
+});
+
+function extractColors() {
+  const elements = document.querySelectorAll("*");
+  const colors = new Set();
+
+  for (let i = 0; i < elements.length; i++) {
+    const element = elements[i];
+    const computedStyle = getComputedStyle(element);
+    const color = computedStyle.color;
+    const backgroundColor = computedStyle.backgroundColor;
+
+    if (
+      color &&
+      color !== "rgba(0, 0, 0, 0)" &&
+      color !== "rgb(0, 0, 0)" &&
+      !colors.has(color)
+    ) {
+      colors.add(color);
+    }
+
+    if (
+      backgroundColor &&
+      backgroundColor !== "rgba(0, 0, 0, 0)" &&
+      backgroundColor !== "rgb(0, 0, 0)" &&
+      !colors.has(backgroundColor)
+    ) {
+      colors.add(backgroundColor);
+    }
+  }
+
+  return Array.from(colors);
+}
+
+function displayColors(colors) {
+  if (!colors.length) return;
+  //Create Rectangles with All Colors from Website
+  colorsAll.innerHTML = colors
+    .map(
+      (color) =>
+        `<li class="color allColorRect">
+          <span class="rect" data-color="${color}" style="background: ${color}; border: 1px solid ${
+          color == "rgb(255, 255, 255)" ? "#ccc" : color
+        }""></span></li>`
+    )
+    .join("");
+
+  document.querySelectorAll(".allColorRect").forEach((li) => {
+    li.addEventListener("click", (e) => {
+      copyColorRect(e.currentTarget.lastElementChild);
+      copiedHeader.classList.remove("hide");
+      setTimeout(() => copiedHeader.classList.add("hide"), 1000);
+    });
+  });
+}
+
+//Get all Colors of the Website in an Array
+
+getAllBtn.addEventListener("click", () => {
+  colorsAllTitle.classList.remove("hide");
+  document.querySelector(".allColorRect").classList.remove("hide");
+});